started styling gallery and made modal
