+++ circles/components/calendar.jsx
@@ -1,47 +1,97 @@
-import { Fragment, useEffect, useRef } from "react";
+import { Fragment } from 'react'
 import {
   ChevronDownIcon,
   ChevronLeftIcon,
   ChevronRightIcon,
+  ClockIcon,
   EllipsisHorizontalIcon,
-} from "@heroicons/react/20/solid";
-import { Menu, Transition } from "@headlessui/react";
-import { days } from "./calendarDates";
+} from '@heroicons/react/20/solid'
+import { Menu, Transition } from '@headlessui/react'
+
+const days = [
+  { date: '2021-12-27', events: [] },
+  { date: '2021-12-28', events: [] },
+  { date: '2021-12-29', events: [] },
+  { date: '2021-12-30', events: [] },
+  { date: '2021-12-31', events: [] },
+  { date: '2022-01-01', isCurrentMonth: true, events: [] },
+  { date: '2022-01-02', isCurrentMonth: true, events: [] },
+  {
+    date: '2022-01-03',
+    isCurrentMonth: true,
+    events: [
+      { id: 1, name: 'Design review', time: '10AM', datetime: '2022-01-03T10:00', href: '#' },
+      { id: 2, name: 'Sales meeting', time: '2PM', datetime: '2022-01-03T14:00', href: '#' },
+    ],
+  },
+  { date: '2022-01-04', isCurrentMonth: true, events: [] },
+  { date: '2022-01-05', isCurrentMonth: true, events: [] },
+  { date: '2022-01-06', isCurrentMonth: true, events: [] },
+  {
+    date: '2022-01-07',
+    isCurrentMonth: true,
+    events: [{ id: 3, name: 'Date night', time: '6PM', datetime: '2022-01-08T18:00', href: '#' }],
+  },
+  { date: '2022-01-08', isCurrentMonth: true, events: [] },
+  { date: '2022-01-09', isCurrentMonth: true, events: [] },
+  { date: '2022-01-10', isCurrentMonth: true, events: [] },
+  { date: '2022-01-11', isCurrentMonth: true, events: [] },
+  {
+    date: '2022-01-12',
+    isCurrentMonth: true,
+    isToday: true,
+    events: [{ id: 6, name: "Sam's birthday party", time: '2PM', datetime: '2022-01-25T14:00', href: '#' }],
+  },
+  { date: '2022-01-13', isCurrentMonth: true, events: [] },
+  { date: '2022-01-14', isCurrentMonth: true, events: [] },
+  { date: '2022-01-15', isCurrentMonth: true, events: [] },
+  { date: '2022-01-16', isCurrentMonth: true, events: [] },
+  { date: '2022-01-17', isCurrentMonth: true, events: [] },
+  { date: '2022-01-18', isCurrentMonth: true, events: [] },
+  { date: '2022-01-19', isCurrentMonth: true, events: [] },
+  { date: '2022-01-20', isCurrentMonth: true, events: [] },
+  { date: '2022-01-21', isCurrentMonth: true, events: [] },
+  {
+    date: '2022-01-22',
+    isCurrentMonth: true,
+    isSelected: true,
+    events: [
+      { id: 4, name: 'Maple syrup museum', time: '3PM', datetime: '2022-01-22T15:00', href: '#' },
+      { id: 5, name: 'Hockey game', time: '7PM', datetime: '2022-01-22T19:00', href: '#' },
+    ],
+  },
+  { date: '2022-01-23', isCurrentMonth: true, events: [] },
+  { date: '2022-01-24', isCurrentMonth: true, events: [] },
+  { date: '2022-01-25', isCurrentMonth: true, events: [] },
+  { date: '2022-01-26', isCurrentMonth: true, events: [] },
+  { date: '2022-01-27', isCurrentMonth: true, events: [] },
+  { date: '2022-01-28', isCurrentMonth: true, events: [] },
+  { date: '2022-01-29', isCurrentMonth: true, events: [] },
+  { date: '2022-01-30', isCurrentMonth: true, events: [] },
+  { date: '2022-01-31', isCurrentMonth: true, events: [] },
+  { date: '2022-02-01', events: [] },
+  { date: '2022-02-02', events: [] },
+  {
+    date: '2022-02-03',
+    events: [{ id: 7, name: 'Cinema with friends', time: '9PM', datetime: '2022-02-04T21:00', href: '#' }],
+  },
+  { date: '2022-02-04', events: [] },
+  { date: '2022-02-05', events: [] },
+  { date: '2022-02-06', events: [] },
+]
+const selectedDay = days.find((day) => day.isSelected)
 
 function classNames(...classes) {
-  return classes.filter(Boolean).join(" ");
+  return classes.filter(Boolean).join(' ')
 }
 
 export default function Example() {
-  const container = useRef(null);
-  const containerNav = useRef(null);
-  const containerOffset = useRef(null);
-
-  useEffect(() => {
-    // Set the container scroll position based on the current time.
-    const currentMinute = new Date().getHours() * 60;
-    container.current.scrollTop =
-      ((container.current.scrollHeight -
-        containerNav.current.offsetHeight -
-        containerOffset.current.offsetHeight) *
-        currentMinute) /
-      1440;
-  }, []);
-
   return (
-    <div className="flex h-full flex-col">
-      <header className="flex flex-none items-center justify-between border-b border-gray-200 py-4 px-6">
-        <div>
-          <h1 className="text-lg font-semibold leading-6 text-gray-900">
-            <time dateTime="2022-01-22" className="sm:hidden">
-              Jan 22, 2022
-            </time>
-            <time dateTime="2022-01-22" className="hidden sm:inline">
-              January 22, 2022
-            </time>
-          </h1>
-          <p className="mt-1 text-sm text-gray-500">Saturday</p>
-        </div>
+    <div className="lg:flex lg:h-full lg:flex-col">
+      <header className="flex items-center justify-between border-b border-gray-200 py-4 px-6 lg:flex-none">
+        <h1 className="text-lg font-semibold text-gray-900">
+          <time dateTime="2022-01">January 2022</time>
+        </h1>
         <div className="flex items-center">
           <div className="flex items-center rounded-md shadow-sm md:items-stretch">
             <button
@@ -72,11 +122,8 @@ export default function Example() {
                 type="button"
                 className="flex items-center rounded-md border border-gray-300 bg-white py-2 pl-3 pr-2 text-sm font-medium text-gray-700 shadow-sm hover:bg-gray-50"
               >
-                Day view
-                <ChevronDownIcon
-                  className="ml-2 h-5 w-5 text-gray-400"
-                  aria-hidden="true"
-                />
+                Month view
+                <ChevronDownIcon className="ml-2 h-5 w-5 text-gray-400" aria-hidden="true" />
               </Menu.Button>
 
               <Transition
@@ -95,10 +142,8 @@ export default function Example() {
                         <a
                           href="#"
                           className={classNames(
-                            active
-                              ? "bg-gray-100 text-gray-900"
-                              : "text-gray-700",
-                            "block px-4 py-2 text-sm"
+                            active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                            'block px-4 py-2 text-sm'
                           )}
                         >
                           Day view
@@ -110,10 +155,8 @@ export default function Example() {
                         <a
                           href="#"
                           className={classNames(
-                            active
-                              ? "bg-gray-100 text-gray-900"
-                              : "text-gray-700",
-                            "block px-4 py-2 text-sm"
+                            active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                            'block px-4 py-2 text-sm'
                           )}
                         >
                           Week view
@@ -125,10 +168,8 @@ export default function Example() {
                         <a
                           href="#"
                           className={classNames(
-                            active
-                              ? "bg-gray-100 text-gray-900"
-                              : "text-gray-700",
-                            "block px-4 py-2 text-sm"
+                            active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                            'block px-4 py-2 text-sm'
                           )}
                         >
                           Month view
@@ -140,10 +181,8 @@ export default function Example() {
                         <a
                           href="#"
                           className={classNames(
-                            active
-                              ? "bg-gray-100 text-gray-900"
-                              : "text-gray-700",
-                            "block px-4 py-2 text-sm"
+                            active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                            'block px-4 py-2 text-sm'
                           )}
                         >
                           Year view
@@ -184,10 +223,8 @@ export default function Example() {
                       <a
                         href="#"
                         className={classNames(
-                          active
-                            ? "bg-gray-100 text-gray-900"
-                            : "text-gray-700",
-                          "block px-4 py-2 text-sm"
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
                         )}
                       >
                         Create event
@@ -201,10 +238,8 @@ export default function Example() {
                       <a
                         href="#"
                         className={classNames(
-                          active
-                            ? "bg-gray-100 text-gray-900"
-                            : "text-gray-700",
-                          "block px-4 py-2 text-sm"
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
                         )}
                       >
                         Go to today
@@ -218,10 +253,8 @@ export default function Example() {
                       <a
                         href="#"
                         className={classNames(
-                          active
-                            ? "bg-gray-100 text-gray-900"
-                            : "text-gray-700",
-                          "block px-4 py-2 text-sm"
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
                         )}
                       >
                         Day view
@@ -233,10 +266,8 @@ export default function Example() {
                       <a
                         href="#"
                         className={classNames(
-                          active
-                            ? "bg-gray-100 text-gray-900"
-                            : "text-gray-700",
-                          "block px-4 py-2 text-sm"
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
                         )}
                       >
                         Week view
@@ -248,10 +279,8 @@ export default function Example() {
                       <a
                         href="#"
                         className={classNames(
-                          active
-                            ? "bg-gray-100 text-gray-900"
-                            : "text-gray-700",
-                          "block px-4 py-2 text-sm"
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
                         )}
                       >
                         Month view
@@ -263,10 +292,8 @@ export default function Example() {
                       <a
                         href="#"
                         className={classNames(
-                          active
-                            ? "bg-gray-100 text-gray-900"
-                            : "text-gray-700",
-                          "block px-4 py-2 text-sm"
+                          active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                          'block px-4 py-2 text-sm'
                         )}
                       >
                         Year view
@@ -279,364 +306,135 @@ export default function Example() {
           </Menu>
         </div>
       </header>
-      <div className="isolate flex flex-auto overflow-hidden bg-white">
-        <div ref={container} className="flex flex-auto flex-col overflow-auto">
-          <div
-            ref={containerNav}
-            className="sticky top-0 z-10 grid flex-none grid-cols-7 bg-white text-xs text-gray-500 shadow ring-1 ring-black ring-opacity-5 md:hidden"
-          >
-            <button
-              type="button"
-              className="flex flex-col items-center pt-3 pb-1.5"
-            >
-              <span>W</span>
-              {/* Default: "text-gray-900", Selected: "bg-gray-900 text-white", Today (Not Selected): "text-indigo-600", Today (Selected): "bg-indigo-600 text-white" */}
-              <span className="mt-3 flex h-8 w-8 items-center justify-center rounded-full text-base font-semibold text-gray-900">
-                19
-              </span>
-            </button>
-            <button
-              type="button"
-              className="flex flex-col items-center pt-3 pb-1.5"
-            >
-              <span>T</span>
-              <span className="mt-3 flex h-8 w-8 items-center justify-center rounded-full text-base font-semibold text-indigo-600">
-                20
-              </span>
-            </button>
-            <button
-              type="button"
-              className="flex flex-col items-center pt-3 pb-1.5"
-            >
-              <span>F</span>
-              <span className="mt-3 flex h-8 w-8 items-center justify-center rounded-full text-base font-semibold text-gray-900">
-                21
-              </span>
-            </button>
-            <button
-              type="button"
-              className="flex flex-col items-center pt-3 pb-1.5"
-            >
-              <span>S</span>
-              <span className="mt-3 flex h-8 w-8 items-center justify-center rounded-full bg-gray-900 text-base font-semibold text-white">
-                22
-              </span>
-            </button>
-            <button
-              type="button"
-              className="flex flex-col items-center pt-3 pb-1.5"
-            >
-              <span>S</span>
-              <span className="mt-3 flex h-8 w-8 items-center justify-center rounded-full text-base font-semibold text-gray-900">
-                23
-              </span>
-            </button>
-            <button
-              type="button"
-              className="flex flex-col items-center pt-3 pb-1.5"
-            >
-              <span>M</span>
-              <span className="mt-3 flex h-8 w-8 items-center justify-center rounded-full text-base font-semibold text-gray-900">
-                24
-              </span>
-            </button>
-            <button
-              type="button"
-              className="flex flex-col items-center pt-3 pb-1.5"
-            >
-              <span>T</span>
-              <span className="mt-3 flex h-8 w-8 items-center justify-center rounded-full text-base font-semibold text-gray-900">
-                25
-              </span>
-            </button>
+      <div className="shadow ring-1 ring-black ring-opacity-5 lg:flex lg:flex-auto lg:flex-col">
+        <div className="grid grid-cols-7 gap-px border-b border-gray-300 bg-gray-200 text-center text-xs font-semibold leading-6 text-gray-700 lg:flex-none">
+          <div className="bg-white py-2">
+            M<span className="sr-only sm:not-sr-only">on</span>
           </div>
-          <div className="flex w-full flex-auto">
-            <div className="w-14 flex-none bg-white ring-1 ring-gray-100" />
-            <div className="grid flex-auto grid-cols-1 grid-rows-1">
-              {/* Horizontal lines */}
-              <div
-                className="col-start-1 col-end-2 row-start-1 grid divide-y divide-gray-100"
-                style={{ gridTemplateRows: "repeat(48, minmax(3.5rem, 1fr))" }}
-              >
-                <div ref={containerOffset} className="row-end-1 h-7"></div>
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    12AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    1AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    2AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    3AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    4AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    5AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    6AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    7AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    8AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    9AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    10AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    11AM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    12PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    1PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    2PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    3PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    4PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    5PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    6PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    7PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    8PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    9PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    10PM
-                  </div>
-                </div>
-                <div />
-                <div>
-                  <div className="sticky left-0 -mt-2.5 -ml-14 w-14 pr-2 text-right text-xs leading-5 text-gray-400">
-                    11PM
-                  </div>
-                </div>
-                <div />
-              </div>
-
-              {/* Events */}
-              <ol
-                className="col-start-1 col-end-2 row-start-1 grid grid-cols-1"
-                style={{
-                  gridTemplateRows: "1.75rem repeat(288, minmax(0, 1fr)) auto",
-                }}
-              >
-                <li
-                  className="relative mt-px flex"
-                  style={{ gridRow: "74 / span 12" }}
-                >
-                  <a
-                    href="#"
-                    className="group absolute inset-1 flex flex-col overflow-y-auto rounded-lg bg-blue-50 p-2 text-xs leading-5 hover:bg-blue-100"
-                  >
-                    <p className="order-1 font-semibold text-blue-700">
-                      Breakfast
-                    </p>
-                    <p className="text-blue-500 group-hover:text-blue-700">
-                      <time dateTime="2022-01-22T06:00">6:00 AM</time>
-                    </p>
-                  </a>
-                </li>
-                <li
-                  className="relative mt-px flex"
-                  style={{ gridRow: "92 / span 30" }}
-                >
-                  <a
-                    href="#"
-                    className="group absolute inset-1 flex flex-col overflow-y-auto rounded-lg bg-pink-50 p-2 text-xs leading-5 hover:bg-pink-100"
-                  >
-                    <p className="order-1 font-semibold text-pink-700">
-                      Flight to Paris
-                    </p>
-                    <p className="order-1 text-pink-500 group-hover:text-pink-700">
-                      John F. Kennedy International Airport
-                    </p>
-                    <p className="text-pink-500 group-hover:text-pink-700">
-                      <time dateTime="2022-01-22T07:30">7:30 AM</time>
-                    </p>
-                  </a>
-                </li>
-                <li
-                  className="relative mt-px flex"
-                  style={{ gridRow: "134 / span 18" }}
-                >
-                  <a
-                    href="#"
-                    className="group absolute inset-1 flex flex-col overflow-y-auto rounded-lg bg-indigo-50 p-2 text-xs leading-5 hover:bg-indigo-100"
-                  >
-                    <p className="order-1 font-semibold text-indigo-700">
-                      Sightseeing
-                    </p>
-                    <p className="order-1 text-indigo-500 group-hover:text-indigo-700">
-                      Eiffel Tower
-                    </p>
-                    <p className="text-indigo-500 group-hover:text-indigo-700">
-                      <time dateTime="2022-01-22T11:00">11:00 AM</time>
-                    </p>
-                  </a>
-                </li>
-              </ol>
-            </div>
+          <div className="bg-white py-2">
+            T<span className="sr-only sm:not-sr-only">ue</span>
+          </div>
+          <div className="bg-white py-2">
+            W<span className="sr-only sm:not-sr-only">ed</span>
+          </div>
+          <div className="bg-white py-2">
+            T<span className="sr-only sm:not-sr-only">hu</span>
+          </div>
+          <div className="bg-white py-2">
+            F<span className="sr-only sm:not-sr-only">ri</span>
+          </div>
+          <div className="bg-white py-2">
+            S<span className="sr-only sm:not-sr-only">at</span>
+          </div>
+          <div className="bg-white py-2">
+            S<span className="sr-only sm:not-sr-only">un</span>
           </div>
         </div>
-        <div className="hidden w-1/2 max-w-md flex-none border-l border-gray-100 py-10 px-8 md:block">
-          <div className="flex items-center text-center text-gray-900">
-            <button
-              type="button"
-              className="-m-1.5 flex flex-none items-center justify-center p-1.5 text-gray-400 hover:text-gray-500"
-            >
-              <span className="sr-only">Previous month</span>
-              <ChevronLeftIcon className="h-5 w-5" aria-hidden="true" />
-            </button>
-            <div className="flex-auto font-semibold">January 2022</div>
-            <button
-              type="button"
-              className="-m-1.5 flex flex-none items-center justify-center p-1.5 text-gray-400 hover:text-gray-500"
-            >
-              <span className="sr-only">Next month</span>
-              <ChevronRightIcon className="h-5 w-5" aria-hidden="true" />
-            </button>
+        <div className="flex bg-gray-200 text-xs leading-6 text-gray-700 lg:flex-auto">
+          <div className="hidden w-full lg:grid lg:grid-cols-7 lg:grid-rows-6 lg:gap-px">
+            {days.map((day) => (
+              <div
+                key={day.date}
+                className={classNames(
+                  day.isCurrentMonth ? 'bg-white' : 'bg-gray-50 text-gray-500',
+                  'relative py-2 px-3'
+                )}
+              >
+                <time
+                  dateTime={day.date}
+                  className={
+                    day.isToday
+                      ? 'flex h-6 w-6 items-center justify-center rounded-full bg-indigo-600 font-semibold text-white'
+                      : undefined
+                  }
+                >
+                  {day.date.split('-').pop().replace(/^0/, '')}
+                </time>
+                {day.events.length > 0 && (
+                  <ol className="mt-2">
+                    {day.events.slice(0, 2).map((event) => (
+                      <li key={event.id}>
+                        <a href={event.href} className="group flex">
+                          <p className="flex-auto truncate font-medium text-gray-900 group-hover:text-indigo-600">
+                            {event.name}
+                          </p>
+                          <time
+                            dateTime={event.datetime}
+                            className="ml-3 hidden flex-none text-gray-500 group-hover:text-indigo-600 xl:block"
+                          >
+                            {event.time}
+                          </time>
+                        </a>
+                      </li>
+                    ))}
+                    {day.events.length > 2 && <li className="text-gray-500">+ {day.events.length - 2} more</li>}
+                  </ol>
+                )}
+              </div>
+            ))}
           </div>
-          <div className="mt-6 grid grid-cols-7 text-center text-xs leading-6 text-gray-500">
-            <div>M</div>
-            <div>T</div>
-            <div>W</div>
-            <div>T</div>
-            <div>F</div>
-            <div>S</div>
-            <div>S</div>
-          </div>
-          <div className="isolate mt-2 grid grid-cols-7 gap-px rounded-lg bg-gray-200 text-sm shadow ring-1 ring-gray-200">
-            {days.map((day, dayIdx) => (
+          <div className="isolate grid w-full grid-cols-7 grid-rows-6 gap-px lg:hidden">
+            {days.map((day) => (
               <button
                 key={day.date}
                 type="button"
                 className={classNames(
-                  "py-1.5 hover:bg-gray-100 focus:z-10",
-                  day.isCurrentMonth ? "bg-white" : "bg-gray-50",
-                  (day.isSelected || day.isToday) && "font-semibold",
-                  day.isSelected && "text-white",
-                  !day.isSelected &&
-                    day.isCurrentMonth &&
-                    !day.isToday &&
-                    "text-gray-900",
-                  !day.isSelected &&
-                    !day.isCurrentMonth &&
-                    !day.isToday &&
-                    "text-gray-400",
-                  day.isToday && !day.isSelected && "text-indigo-600",
-                  dayIdx === 0 && "rounded-tl-lg",
-                  dayIdx === 6 && "rounded-tr-lg",
-                  dayIdx === days.length - 7 && "rounded-bl-lg",
-                  dayIdx === days.length - 1 && "rounded-br-lg"
+                  day.isCurrentMonth ? 'bg-white' : 'bg-gray-50',
+                  (day.isSelected || day.isToday) && 'font-semibold',
+                  day.isSelected && 'text-white',
+                  !day.isSelected && day.isToday && 'text-indigo-600',
+                  !day.isSelected && day.isCurrentMonth && !day.isToday && 'text-gray-900',
+                  !day.isSelected && !day.isCurrentMonth && !day.isToday && 'text-gray-500',
+                  'flex h-14 flex-col py-2 px-3 hover:bg-gray-100 focus:z-10'
                 )}
               >
                 <time
                   dateTime={day.date}
                   className={classNames(
-                    "mx-auto flex h-7 w-7 items-center justify-center rounded-full",
-                    day.isSelected && day.isToday && "bg-indigo-600",
-                    day.isSelected && !day.isToday && "bg-gray-900"
+                    day.isSelected && 'flex h-6 w-6 items-center justify-center rounded-full',
+                    day.isSelected && day.isToday && 'bg-indigo-600',
+                    day.isSelected && !day.isToday && 'bg-gray-900',
+                    'ml-auto'
                   )}
                 >
-                  {day.date.split("-").pop().replace(/^0/, "")}
+                  {day.date.split('-').pop().replace(/^0/, '')}
                 </time>
+                <span className="sr-only">{day.events.length} events</span>
+                {day.events.length > 0 && (
+                  <span className="-mx-0.5 mt-auto flex flex-wrap-reverse">
+                    {day.events.map((event) => (
+                      <span key={event.id} className="mx-0.5 mb-1 h-1.5 w-1.5 rounded-full bg-gray-400" />
+                    ))}
+                  </span>
+                )}
               </button>
             ))}
           </div>
         </div>
       </div>
+      {selectedDay?.events.length > 0 && (
+        <div className="py-10 px-4 sm:px-6 lg:hidden">
+          <ol className="divide-y divide-gray-100 overflow-hidden rounded-lg bg-white text-sm shadow ring-1 ring-black ring-opacity-5">
+            {selectedDay.events.map((event) => (
+              <li key={event.id} className="group flex p-4 pr-6 focus-within:bg-gray-50 hover:bg-gray-50">
+                <div className="flex-auto">
+                  <p className="font-semibold text-gray-900">{event.name}</p>
+                  <time dateTime={event.datetime} className="mt-2 flex items-center text-gray-700">
+                    <ClockIcon className="mr-2 h-5 w-5 text-gray-400" aria-hidden="true" />
+                    {event.time}
+                  </time>
+                </div>
+                <a
+                  href={event.href}
+                  className="ml-6 flex-none self-center rounded-md border border-gray-300 bg-white py-2 px-3 font-semibold text-gray-700 opacity-0 shadow-sm hover:bg-gray-50 focus:opacity-100 group-hover:opacity-100"
+                >
+                  Edit<span className="sr-only">, {event.name}</span>
+                </a>
+              </li>
+            ))}
+          </ol>
+        </div>
+      )}
     </div>
-  );
+  )
 }
