--- circles/components/events-data.js
@@ -1,47 +1,0 @@
-export const data = [
-    {
-        id: 1,
-        image: "https://imengine.prod.srp.navigacloud.com/?uuid=f477ce59-bcd1-5089-a42e-f57b058953b2&type=primary&q=72&width=1024",
-        name: "Chess Club",
-        description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
-        time: "tuesdays 6pm to 8pm",
-        location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
-        availability: "Available"
-    },
-    {
-        id: 2,
-        image: "https://swimclubmanager.blob.core.windows.net/files-8201/blobid1528032636707.jpg",
-        name: "Swimming Club",
-        description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
-        time: "tuesdays 6pm to 8pm",
-        location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
-        availability: "Available"
-    },
-    {
-        id: 3,
-        image: "https://www.stamfordschools.org.uk/wp-content/uploads/2020/01/climbing_0004_3.jpg",
-        name: "Rock Climbing Club",
-        description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
-        time: "tuesdays 6pm to 8pm",
-        location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
-        availability: "Available"
-    },
-    {
-        id: 5,
-        image: "https://www.stamfordschools.org.uk/wp-content/uploads/2020/01/climbing_0004_3.jpg",
-        name: "Rock Climbing Club",
-        description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
-        time: "tuesdays 6pm to 8pm",
-        location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
-        availability: "Available"
-    },
-    {
-        id: 6,
-        image: "https://www.stamfordschools.org.uk/wp-content/uploads/2020/01/climbing_0004_3.jpg",
-        name: "Rock Climbing Club",
-        description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
-        time: "tuesdays 6pm to 8pm",
-        location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
-        availability: "Available"
-    }
-]
