+++ circles/data/events-data.js
@@ -1,0 +1,57 @@
+export const data = [
+  {
+    id: 1,
+    name: "Ageing Well Service from Age UK Somerset",
+    image:
+      "https://media.istockphoto.com/photos/group-of-senior-retirement-friends-happiness-concept-picture-id660654876?b=1&k=20&m=660654876&s=612x612&w=0&h=-TjWkwkGeVwiMlKaUQvqLkWbh3N5nqWT8gFiYcsRtLg=",
+    tagline:
+      "If you’re an older adult looking to join an exercise class, then why not try an Ageing Well class with Age UK Somerset?",
+    description:
+      "Physical activity helps you stay healthy, active and independent as you get older and it’s not all about high impact exercise – it’s about finding something that’s right for you. Age UK Somerset is dedicated to promoting the health and wellbeing of older people in North Somerset and Somerset.",
+    time: "Mon-Fri 9am to 5pm",
+    location: "Weston-super-Mare, BS23, United Kingdom",
+    availability: "Available",
+    email: "ageingwell@ageuksomerset.org.uk",
+    telephone: "01823 345624",
+  },
+  {
+    id: 2,
+    image:
+      "https://swimclubmanager.blob.core.windows.net/files-8201/blobid1528032636707.jpg",
+    name: "Swimming Club",
+    description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
+    time: "tuesdays 6pm to 8pm",
+    location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
+    availability: "Available",
+  },
+  {
+    id: 3,
+    image:
+      "https://www.stamfordschools.org.uk/wp-content/uploads/2020/01/climbing_0004_3.jpg",
+    name: "Rock Climbing Club",
+    description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
+    time: "tuesdays 6pm to 8pm",
+    location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
+    availability: "Available",
+  },
+  {
+    id: 5,
+    image:
+      "https://www.stamfordschools.org.uk/wp-content/uploads/2020/01/climbing_0004_3.jpg",
+    name: "Rock Climbing Club",
+    description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
+    time: "tuesdays 6pm to 8pm",
+    location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
+    availability: "Available",
+  },
+  {
+    id: 6,
+    image:
+      "https://www.stamfordschools.org.uk/wp-content/uploads/2020/01/climbing_0004_3.jpg",
+    name: "Rock Climbing Club",
+    description: "good club adajfbn  bfai  frf ; faeubar f auref  ",
+    time: "tuesdays 6pm to 8pm",
+    location: "Laurel Rd, Handsworth, Birmingham B21 9PB",
+    availability: "Available",
+  },
+];
