+++ circles/pages/_app.js
@@ -1,12 +1,11 @@
-import '../styles/globals.css'
-import Navbar from '../components/navbar.jsx'
+import "../styles/globals.css";
+import Navbar from "../components/Navbar.jsx";
 
 export default function App({ Component, pageProps }) {
-  
   return (
     <>
-    <Navbar/>
-   <Component {...pageProps} />
-   </>
-  )
+      <Navbar />
+      <Component {...pageProps} />
+    </>
+  );
 }
